Split test_runtime_env_working_dir_remote_uri into 3 tests
